Rename useTheme file to use-theme
