+++ root/Card/card-body.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react';
 
 import { twMerge } from 'tailwind-merge';
 
-import { useTheme } from '../themes/useTheme';
+import { useTheme } from '../themes/use-theme';
 import type { DivProps } from '../types';
 
 const CardBody = React.forwardRef<HTMLDivElement, DivProps>((props, ref) => {
