+++ root/Checkbox/Checkbox.tsx
@@ -2,7 +2,7 @@ import { ForwardedRef, forwardRef } from 'react';
 
 import { twMerge } from 'tailwind-merge';
 
-import { useTheme } from '../themes/useTheme';
+import { useTheme } from '../themes/use-theme';
 
 import type { CheckboxProps, CheckboxRef } from './types';
