+++ root/Modal/modal-body.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Dialog } from '@headlessui/react';
 import { twMerge } from 'tailwind-merge';
 
-import { useTheme } from '../themes/useTheme';
+import { useTheme } from '../themes/use-theme';
 
 import type { ModalBodyProps } from './types';
