+++ root/Modal/modal-header.tsx
@@ -1,7 +1,7 @@
 import { Dialog } from '@headlessui/react';
 import { twMerge } from 'tailwind-merge';
 
-import { useTheme } from '../themes/useTheme';
+import { useTheme } from '../themes/use-theme';
 
 import type { ModalHeaderProps } from './types';
