+++ root/Select/use-select.ts
@@ -2,7 +2,7 @@ import { ChangeEvent, useCallback } from 'react';
 
 import { twMerge } from 'tailwind-merge';
 
-import { useTheme } from '../themes/useTheme';
+import { useTheme } from '../themes/use-theme';
 
 import type { ISelectOption, SelectProps } from './types';
