+++ root/Tooltip/Tooltip.tsx
@@ -2,7 +2,7 @@ import { PropsWithChildren } from 'react';
 
 import Tippy from '@tippyjs/react';
 
-import { useTheme } from '../themes/useTheme';
+import { useTheme } from '../themes/use-theme';
 
 import type { TooltipProps } from './types';
