+++ root/Typography/text.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react';
 
 import { twMerge } from 'tailwind-merge';
 
-import { useTheme } from '../themes/useTheme';
+import { useTheme } from '../themes/use-theme';
 
 import type { TextProps } from './types';
